Use DataFrame for metadata instead of dict
Allows metadata to be returned as a pandas DataFrame instead of a
dictionary and uses this API to load metadata as a DataFrame in augur
filter. Updates the corresponding metadata references in augur filter
and related function calls to work with this new data structure.

This change allows us to benefit from pandas's lower memory footprint
and eventually switch to categorical variables for object-typed metadata
columns that can be enumerated and also switch to chunking metadata
input to avoid reading it all into memory.